Update entities
When setting up Rhasspy, I realized there are some entities I didn't want to be voice activated, and I couldn't find a way to prevent them from being added to the entity list, so I modified this script slightly to check if "${RHASSPY_PROFILE_DIR}/slot_programs/hass/excluded_entities" exists, and compile an array of excluded entities from it. 

I created the "excluded_entities" file at the aforementioned location and added the entities I personally wanted to exclude, each as a line item in the file. Then, I expanded the IF check in line 62 to determine if the entity_id could be found in the array of excluded entities. If the entity_id isn't found in the array of exclusions, then it can be added.

This has been tested and is working well in my Rhasspy setup. I hope this small change might be helpful to some folks.